third_party/blink/web_tests/http/tests/media/media-playback-while-not-visible-permission-policy-cant-play-if-hidden-cross-origin.html

<!DOCTYPE html>
<title>Test behavior of cross-origin HTMLMediaElements with the media-playback-while-not-rendered permission policy</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/media-playback-while-not-visible-utils.js"></script>
<body>
<script>
  internals.settings.setAutoplayPolicy('no-user-gesture-required');

  function hideFrame(iframe, type) {
    if (type === 'display') {
      iframe.style.setProperty('display', 'none');
    } else if (type === 'visibility') {
      iframe.style.setProperty('visibility', 'hidden');
    } else if (type === 'zeroSize') {
      iframe.style.setProperty('width', '0');
      iframe.style.setProperty('height', '0');
    }
  }

  let visibility_types = ['display', 'visibility', 'zeroSize'];
  for (let type of visibility_types) {
    promise_test(async t => {
      let iframe = document.createElement('iframe');
      iframe.allow = "media-playback-while-not-visible 'none'; autoplay *";
      iframe.src = "http://localhost:8000/media/resources/media-playback-while-not-visible-iframe.html";
      document.body.appendChild(iframe);
      await new Promise(resolve => {
        iframe.addEventListener('load', resolve);
      });

      t.add_cleanup(() => {
        document.body.removeChild(iframe);
      });

      hideFrame(iframe, type);

      // Wait for the cross-origin iframe visibility state to propagate.
      await new Promise(resolve => setTimeout(resolve, 2000));

      // Attempting to play while hidden should fail with NotAllowedError.
      assert_equals(await playMediaInIframe(iframe), 'NotAllowedError');
      assert_equals(await queryPlayerStatus(iframe), 'Paused');
    }, "Cross-origin media elements in non-rendered iframes shouldn't be allowed to play (type = " + type + ")");
  }
</script>
</body>
